Replace secret `curl` headers with a config file
Putting secret tokens in command-line arguments is dangerous in the case
of command logging (`set -x`) or other processes running on the system
(command lines are publicly visible). The standard fix is to pass tokens
in environment variables or properly protected files. The base image has
curl 7.52 < 7.55, so we can’t use `-H @headersfile`, but we can create a
config file.

wchargin-branch: curl-config